+++ udata/api_fields.py
@@ -70,10 +70,14 @@ def convert_db_to_field(key, field, info = {}):
         constructor_write = restx_fields.String
     elif isinstance(field, mongo_fields.EmbeddedDocumentField):
         nested_fields = info.get('nested_fields')
-        if nested_fields is None:
-            raise ValueError(f"EmbeddedDocumentField `{key}` requires a `nested_fields` param to serialize/deserialize.")
+        if nested_fields is not None:
+            constructor = lambda **kwargs: restx_fields.Nested(nested_fields, **kwargs)
+        elif hasattr(field.document_type_obj, '__read_fields__'):
+            constructor_read = lambda **kwargs: restx_fields.Nested(field.document_type_obj.__read_fields__, **kwargs)
+            constructor_write = lambda **kwargs: restx_fields.Nested(field.document_type_obj.__write_fields__, **kwargs)
+        else:
+            raise ValueError(f"EmbeddedDocumentField `{key}` requires a `nested_fields` param to serialize/deserialize or a `@generate_fields()` definition.")
 
-        constructor = lambda **kwargs: restx_fields.Nested(nested_fields, **kwargs)
     else:
         raise ValueError(f"Unsupported MongoEngine field type {field.__class__.__name__}")
     
@@ -96,6 +100,7 @@ def generate_fields(**kwargs):
         read_fields = {}
         write_fields = {}
         sortables = []
+        filterables = []
 
         read_fields['id'] = restx_fields.String(required=True)
 
@@ -105,6 +110,23 @@ def generate_fields(**kwargs):
 
             if info.get('sortable', False):
                 sortables.append(key)
+
+            filterable = info.get('filterable', None)
+            if filterable is not None:
+                if 'key' not in filterable:
+                    filterable['key'] = key
+                if 'column' not in filterable:
+                    filterable['column'] = key
+
+                if 'constraints' not in filterable:
+                    filterable['constraints'] = []
+                    if isinstance(field, mongo_fields.ReferenceField) or (isinstance(field, mongo_fields.ListField) and isinstance(field.field, mongo_fields.ReferenceField)):
+                        filterable['constraints'].append('objectid')
+
+                # We may add more information later here:
+                # - type of mongo query to execute (right now only simple =)
+
+                filterables.append(filterable)
 
             read, write = convert_db_to_field(key, field)
 
@@ -159,12 +181,25 @@ def generate_fields(**kwargs):
             choices = sortables + ['-' + k for k in sortables]
             parser.add_argument('sort', type=str, location='args', choices=choices, help='The field (and direction) on which sorting apply')
 
+        for filterable in filterables:
+            parser.add_argument(filterable['key'], type=str, location='args')
+
         cls.__index_parser__ = parser
         def apply_sort_filters_and_pagination(base_query):
             args = cls.__index_parser__.parse_args()
 
             if sortables and args['sort']:
                 base_query = base_query.order_by(args['sort'])
+
+            for filterable in filterables:
+                if args.get(filterable['key']):
+                    for constraint in filterable['constraints']:
+                        if constraint == 'objectid' and not ObjectId.is_valid(args[filterable['key']]):
+                            api.abort(400, f'`{filterable["key"]}` must be an identifier')
+
+                    base_query = base_query.filter(**{
+                        filterable['column']: args[filterable['key']],
+                    })
 
             if paginable:
                 base_query = base_query.paginate(args['page'], args['page_size'])