+++ udata/core/dataservices/models.py
@@ -3,11 +3,11 @@ from udata.api_fields import field, function_field, generate_fields
 from udata.core.dataset.models import Dataset
 from udata.core.metrics.models import WithMetrics
 from udata.core.owned import Owned, OwnedQuerySet
-from udata.i18n import lazy_gettext as _
 import udata.core.contact_point.api_fields as contact_api_fields
 import udata.core.dataset.api_fields as datasets_api_fields
+from udata.i18n import lazy_gettext as _
 
-from udata.models import db
+from udata.models import db, Discussion, Follow
 from udata.uris import endpoint_for
 
 # "frequency"
@@ -30,6 +30,35 @@ class DataserviceQuerySet(OwnedQuerySet):
         return self(db.Q(private=True) |
                     db.Q(deleted_at__ne=None) |
                     db.Q(archived_at__ne=None))
+
+@generate_fields()
+class HarvestMetadata(db.EmbeddedDocument):
+    backend = field(db.StringField())
+    domain = field(db.StringField())
+
+    source_id = field(db.StringField())
+    source_url = field(db.URLField())
+
+    remote_id = field(db.StringField())
+    remote_url = field(db.URLField())
+
+    # If the node ID is a `URIRef` it means it links to something external, if it's not an `URIRef` it's often a
+    # auto-generated ID just to link multiple RDF node togethers. When exporting as RDF to other catalogs, we 
+    # want to re-use this node ID (only if it's not auto-generated) to improve compatibility.
+    uri = field(
+        db.URLField(),
+        description="RDF node ID if it's an `URIRef`. `None` if it's not present or if it's a random auto-generated ID inside the graph.",
+    )
+
+    created_at = field(
+        db.DateTimeField(),
+        description="Date of the creation as provided by the harvested catalog"
+    )
+    last_update = field(
+        db.DateTimeField(),
+        description="Date of the last harvesting"
+    )
+    archived_at = field(db.DateTimeField())
 
 @generate_fields()
 class Dataservice(WithMetrics, Owned, db.Document):
@@ -111,20 +140,41 @@ class Dataservice(WithMetrics, Owned, db.Document):
         db.ListField(
             field(
                 db.ReferenceField(Dataset),
-                nested_fields=datasets_api_fields.dataset_fields,
+                nested_fields=datasets_api_fields.dataset_ref_fields,
             )
-        )
+        ),
+        filterable={
+            'key': 'dataset',
+        },
+    )
+
+    harvest = field(
+        db.EmbeddedDocumentField(HarvestMetadata),
+        readonly=True,
     )
 
     @function_field(description="Link to the API endpoint for this dataservice")
     def self_api_url(self):
         return endpoint_for('api.dataservice', dataservice=self, _external=True)
 
-    def self_web_url():
-        pass
+    @function_field(description="Link to the udata web page for this dataservice")
+    def self_web_url(self):
+        return endpoint_for('dataservices.show', dataservice=self, _external=True)
 
     # TODO
     # frequency = db.StringField(choices=list(UPDATE_FREQUENCIES.keys()))
     # temporal_coverage = db.EmbeddedDocumentField(db.DateRange)
     # spatial = db.EmbeddedDocumentField(SpatialCoverage)
     # harvest = db.EmbeddedDocumentField(HarvestDatasetMetadata)
+
+    @property
+    def is_hidden(self):
+        return self.private or self.deleted_at or self.archived_at
+
+    def count_discussions(self):
+        self.metrics['discussions'] = Discussion.objects(subject=self, closed=None).count()
+        self.save()
+
+    def count_followers(self):
+        self.metrics['followers'] = Follow.objects(until=None).followers(self).count()
+        self.save()